Added optional multilingual site `footer_text`
If `multilingual: true` is set in the `_config.yml` file (and one is using the polyglot plugin), the site's `footer_text` can be displayed according to the currently selected language.
In `_config.yml` one can add a `footer_text` for each language, e.g.
```
footer_text:
  en: Powered by <a href="https://jekyllrb.com/">Jekyll</a> with <a href="https://github.com/sylhare/Type-on-Strap">Type on Strap</a>
  de: Webseite mit <a href="https://jekyllrb.com/">Jekyll</a> sowie <a href="https://github.com/sylhare/Type-on-Strap">Type on Strap</a> erstellt
```

If `multilingual: true` is not set, everything will work as before.