Added the capability to change the word order of localized strings
Added the capability to format the strings with "%s" to change the word order
---in languages other than English, this will sometimes be necessary, e.g.
in German, "Auf Facebook folgen" is better grammar than "Folgen auf Facebook";
this can be done now by adding `str_follow_on: "Auf %s folgen"` to the
`language.yml` data file.

If no "%s" is specified in the localization string, the default behaviour of
appending it to the end will be applied.

Also added `str_share` to `language.yml`---so that this can also be localized.